Added first draft of the second report
Merged both first and second reports.
Added all the attachments (Use Cases + Database Tables).
Fixed the indexes.
Moved the conclusion section to the lowest part of the report.
Reformulated some sections as to refer to this second report (such as
the initial summary).

TO DO: The commented sections are yet to be done/edited accordingly (or
just need to be looked at).